Refactor mesher
