+++ src/client/terrain/util/mod.rs
@@ -1,5 +1,9 @@
 pub mod blocks;
 pub mod mesher;
+pub mod cube_mesher;
+pub mod cross_mesher;
 
 pub use blocks::*;
 pub use mesher::*;
+pub use cube_mesher::*;
+pub use cross_mesher::*;
